wip(frontend): fuse actionbar and search area on question pool
